+++ prompts/image_input.txt
