+++ requirements.txt
@@ -1,63 +1,4 @@
-annotated-types==0.7.0
-antlr4-python3-runtime==4.9.3
-anyio==4.11.0
-blinker==1.9.0
-cachetools==6.2.0
-certifi==2025.8.3
-charset-normalizer==3.4.3
-click==8.3.0
-deptry==0.23.1
-docopt==0.6.2
-easyocr==1.7.2
-filelock==3.19.1
-Flask==3.1.2
-fsspec==2025.9.0
-google-auth==2.41.1
-google-genai==1.41.0
-h11==0.16.0
-httpcore==1.0.9
-httpx==0.28.1
-hydra-core==1.3.2
-idna==3.10
-imageio==2.37.0
-itsdangerous==2.2.0
-lazy_loader==0.4
-mpmath==1.3.0
-networkx==3.5
-ninja==1.13.0
-numpy==2.2.6
-omegaconf==2.3.0
-opencv-python==4.12.0.88
-opencv-python-headless==4.12.0.88
-packaging==25.0
-pillow==11.3.0
-pip-autoremove==0.10.0
-pipreqs==0.4.13
-pyasn1==0.6.1
-pyasn1_modules==0.4.2
-pyclipper==1.3.0.post6
-pydantic==2.11.10
-pydantic_core==2.33.2
-python-bidi==0.6.6
+Pillow==11.3.0
+protobuf==6.32.1
 python-dotenv==1.1.1
 python-telegram-bot==22.5
-PyYAML==6.0.3
-requests==2.32.5
-requirements-parser==0.13.0
-rsa==4.9.1
-scikit-image==0.25.2
-scipy==1.16.2
-setuptools==80.9.0
-shapely==2.1.2
-sniffio==1.3.1
-sympy==1.14.0
-tenacity==9.1.2
-tifffile==2025.9.30
-torch==2.8.0
-torchvision==0.23.0
-typing-inspection==0.4.2
-typing_extensions==4.15.0
-urllib3==2.5.0
-websockets==15.0.1
-Werkzeug==3.1.3
-yarg==0.1.10
