+++ src/bot.py
@@ -1,4 +1,5 @@
 from utils.telegram_handlers import photo_handler
+from utils.telegram_handlers import text_handler
 from telegram.ext import Application, MessageHandler, filters
 from config import TOKEN, initialize_directories
 
@@ -11,8 +12,10 @@ def main() -> None:
 
     # Thêm trình xử lý cho tin nhắn ảnh
     application.add_handler(MessageHandler(filters.PHOTO, photo_handler))
+    
     # Trình xử lí cho tin nhắn văn bản
-    #application.add_handler(MessageHandler(filters.TEXT & (~filters.COMMAND), text_handler))
+    application.add_handler(MessageHandler(filters.TEXT & (~filters.COMMAND), text_handler))
+    
     # Bắt đầu chạy bot
     print("Bot đang chạy...")
     application.run_polling()
